+++ test/e2e/e2e_test.go
@@ -187,7 +187,9 @@ metadata:
 
 		It("should ensure the metrics endpoint is serving metrics", func() {
 			By("creating a ClusterRoleBinding for the service account to allow access to metrics")
-			_, _ = utils.Run(exec.Command("kubectl", "delete", "clusterrolebinding", metricsRoleBindingName, "--ignore-not-found"))
+			_, _ = utils.Run(exec.Command(
+				"kubectl", "delete", "clusterrolebinding", metricsRoleBindingName, "--ignore-not-found",
+			))
 			cmd := exec.Command("kubectl", "create", "clusterrolebinding", metricsRoleBindingName,
 				"--clusterrole=project-onboarding-operator-metrics-reader",
 				fmt.Sprintf("--serviceaccount=%s:%s", namespace, serviceAccountName),